+++ trunk/com/planet_ink/coffee_mud/Common/DefaultRideable.java
@@ -9,12 +9,14 @@ import com.planet_ink.coffee_mud.Common.interfaces.*;
 import com.planet_ink.coffee_mud.Exits.interfaces.*;
 import com.planet_ink.coffee_mud.Items.interfaces.*;
 import com.planet_ink.coffee_mud.Locales.interfaces.*;
+import com.planet_ink.coffee_mud.Libraries.interfaces.*;
 import com.planet_ink.coffee_mud.MOBS.interfaces.*;
 import com.planet_ink.coffee_mud.Races.interfaces.*;
 
 import java.util.*;
-import java.io.IOException;
 import java.nio.ByteBuffer;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.io.IOException;
 
 /*
 CoffeeMUD 5.6.2 copyright 2000-2010 Bo Zimmerman
@@ -25,9 +27,10 @@ Licensed under the Apache License, Version 2.0. You may obtain a copy of the lic
 */
 public class DefaultRideable implements Rideable, Ownable
 {
-	protected Vector<Item> riders=new Vector<Item>(1);
+	protected CopyOnWriteArrayList<Item> riders=new CopyOnWriteArrayList();
 	protected CMSavable parent=null;
 	protected int saveNum=0;
+	protected boolean amDestroyed=false;
 	protected boolean mobile=false;
 /*
 	protected String putString="on";
@@ -51,7 +54,13 @@ public class DefaultRideable implements Rideable, Ownable
 	public void destroy()
 	{
 		//TODO
-		CMLib.database().deleteObject(this);
+		amDestroyed=true;
+		if(saveNum!=0)
+			CMLib.database().deleteObject(this);
+	}
+	public boolean amDestroyed()
+	{
+		return amDestroyed;
 	}
 
 	//CMModifiable and CMSavable
@@ -61,11 +70,11 @@ public class DefaultRideable implements Rideable, Ownable
 	public Enum[] headerEnumM(){return new Enum[] {MCode.values()[0]};}
 	public int saveNum()
 	{
-		if(saveNum==0)
+		if((saveNum==0)&&(!amDestroyed))
 		synchronized(this)
 		{
 			if(saveNum==0)
-				saveNum=SIDLib.Objects.RIDEABLE.getNumber(this);
+				saveNum=SIDLib.RIDEABLE.getNumber(this);
 		}
 		return saveNum;
 	}
@@ -74,14 +83,15 @@ public class DefaultRideable implements Rideable, Ownable
 		synchronized(this)
 		{
 			if(saveNum!=0)
-				SIDLib.Objects.RIDEABLE.removeNumber(saveNum);
+				SIDLib.RIDEABLE.removeNumber(saveNum);
 			saveNum=num;
-			SIDLib.Objects.RIDEABLE.assignNumber(num, this);
+			SIDLib.RIDEABLE.assignNumber(num, this);
 		}
 	}
 	public boolean needLink(){return false;}
 	public void link(){}
 	public void saveThis(){CMLib.database().saveObject(this);}
+	public void prepDefault(){}
 
 	//Rideable
 	public boolean isMobileRide(){return mobile;}
@@ -90,13 +100,13 @@ public class DefaultRideable implements Rideable, Ownable
 	public boolean hasRider(Item E) { return riders.contains(E); }
 	public void addRider(Item E)
 	{
-		E.setRide(this);
-		riders.addElement(E);
+		E.setRide(parent);
+		riders.add(E);
 //		E.getEnvObject().recoverEnvStats();
 	}
 	public void removeRider(Item E)
 	{
-		riders.removeElement(E);
+		riders.remove(E);
 		E.setRide(null);
 //		E.getEnvObject().recoverEnvStats();
 	}
@@ -109,11 +119,11 @@ public class DefaultRideable implements Rideable, Ownable
 	}
 	public Item getRider(int index)
 	{
-		try { return riders.elementAt(index); }
+		try { return riders.get(index); }
 		catch(java.lang.ArrayIndexOutOfBoundsException x){}
 		return null;
 	}
-	public Vector<Item> allRiders(){return (Vector<Item>)riders.clone();}
+	public Iterator<Item> allRiders(){return riders.iterator();}
 	public int numRiders() { return riders.size(); }
 	public String putString(Item R) { return "on"; }
 	public String stateString(Item R) { return "sitting on"; }
@@ -144,7 +154,7 @@ public class DefaultRideable implements Rideable, Ownable
 				boolean done=false;
 				while((M.session()!=null)&&(!M.session().killFlag())&&(!done))
 				{
-					Vector<Item> V=E.allRiders();
+					Vector<Item> V=new Vector(E.riders);
 					int i=CMLib.genEd().promptVector(M, V, false);
 					if(--i<0) done=true;
 					else if(i<V.size())
